+++ packages/theme-patternfly-org/global.css
@@ -1,0 +1,3 @@
+#root {
+  height: 100%;
+}
