+++ packages/theme-patternfly-org/helpers/extractExamples.js
@@ -1,17 +1,6 @@
 const visit = require('unist-util-visit');
 const { render } = require('html-formatter');
 const { getId } = require('./getId');
-const { slugger } = require('./slugger');
-
-// Used in core's dev:lite
-function getExampleClassName(source, componentType, componentName) {
-  return `ws-${source}-${componentType}-${componentName}`;
-}
-
-// Used in core's dev:lite
-function getExampleId(source, componentType, componentName, exampleTitle) {
-  return `ws-${source}-${componentType}-${componentName}-${slugger(exampleTitle)}`;
-}
 
 module.exports = {
   // Map example page urls to HTML
@@ -87,9 +76,5 @@ module.exports = {
     });
 
     return examples;
-  },
-  getExampleClassName,
-  getExampleId
+  }
 }
-
-// TODO: Write some tests for example MDXAsts
