+++ packages/theme-patternfly-org/package.json
@@ -5,12 +5,34 @@
   "author": "Red Hat",
   "license": "MIT",
   "dependencies": {
+    "@reach/router": "^1.3.4",
     "codesandbox": "^2.1.10",
     "file-saver": "^1.3.8",
     "html-formatter": "0.1.9",
-    "null-loader": "^3.0.0",
     "prismjs": "^1.17.1",
     "react-helmet": "^5.2.1",
     "react-live": "^2.2.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.10.5",
+    "@babel/plugin-proposal-class-properties": "^7.10.4",
+    "@babel/plugin-transform-react-jsx": "^7.10.4",
+    "@babel/preset-env": "^7.10.4",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.3.3",
+    "babel-loader": "^8.1.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^6.0.3",
+    "css-loader": "^3.6.0",
+    "file-loader": "^6.0.0",
+    "html-webpack-plugin": "^4.3.0",
+    "mini-css-extract-plugin": "^0.9.0",
+    "null-loader": "^3.0.0",
+    "react-refresh": "^0.8.3",
+    "size-plugin": "^2.0.1",
+    "url-loader": "^4.1.0",
+    "webpack": "^4.43.0",
+    "webpack-bundle-analyzer": "^3.8.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
   }
 }
