+++ packages/v4/patternfly-docs.config.js
@@ -1,0 +1,19 @@
+module.exports = {
+  hasGdprBanner: true,
+  hasFooter: true,
+  hasSearch: true,
+  hasVersionSwitcher: true,
+  sideNavItems: [
+    { section: 'get-started' },
+    { section: 'styles' },
+    { section: 'content' },
+    { section: 'components' },
+    { section: 'demos' },
+    { section: 'layouts' },
+    { section: 'contribute' },
+    { href: '/get-in-touch' },
+  ],
+  topNavItems: [],
+  prnum: null,
+  prurl: null
+};
