+++ packages/v4/patternfly-docs.css.js
@@ -1,0 +1,12 @@
+// React-specific CSS
+import '@patternfly/react-styles/src/css/components/Table/inline-edit.css'; // inline-edit extension styles
+import '@patternfly/react-styles/src/css/components/Topology/topology-controlbar.css';
+import '@patternfly/react-styles/src/css/components/Topology/topology-side-bar.css';
+import '@patternfly/react-styles/src/css/components/Topology/topology-view.css';
+import '@patternfly/react-styles/src/css/layouts/Toolbar/toolbar.css';
+// Global theme CSS
+import 'theme-patternfly-org/global.css';
+// Patternfly
+import '@patternfly/patternfly/patternfly.css';
+// Utilities
+import '@patternfly/patternfly/patternfly-addons.css';
