+++ packages/v4/src/app.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { Router } from '@reach/router';
+import { HomePage, GetInTouchPage } from './pages';
+import * as LayoutOptions from '../patternfly-docs.config.js';
+import '../patternfly-docs.css.js';
+
+function App() {
+  return (
+    <Router>
+      <HomePage path="/" layoutOptions={LayoutOptions} />
+      <GetInTouchPage path="/get-in-touch" layoutOptions={LayoutOptions} />
+    </Router>
+  );
+}
+
+ReactDOM.render(<App />, document.getElementById('root'));
