+++ packages/v4/src/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <head>
+    <meta charset="utf-8">
+    <title>PatternFly 4</title>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
+</html>
