+++ packages/v4/webpack.config.js
@@ -1,0 +1,1 @@
+module.exports = require('theme-patternfly-org/webpack.config');
